Scaling Sprite Added! Fixed bunches of width and height design mistake

+++ SuperDavis/SuperDavis/Interfaces/IDavisState.cs
@@ -2,8 +2,8 @@
 {
     interface IDavisState : IGameObjectState
     {
-        new int Width { get; set; }
-        new int Height { get; set; }
+        new float Width { get; set; }
+        new float Height { get; set; }
         void Static();
         void Left();
         void Right();
