+++ SuperDavis/SuperDavis/Interfaces/IGameState.cs
@@ -5,8 +5,8 @@ namespace SuperDavis.Interfaces
 {
     interface IGameObjectState
     {
-        int Width { get; set; }
-        int Height { get; set; }
+        float Width { get; set; }
+        float Height { get; set; }
         ISprite Sprite { get; set; }
         void Update(GameTime gameTime);
         void Draw(SpriteBatch spriteBatch, Vector2 location);
