+++ SuperDavis/SuperDavis/Object/Block/Pipe.cs
@@ -21,7 +21,7 @@ namespace SuperDavis.Object.Block
             Location = location;
             pipeStateMachine = new PipeStateMachine();
             block = pipeStateMachine.Sprite;
-            HitBox = new Rectangle((int)Location.X, (int)Location.Y, block.Width, block.Height);
+            HitBox = new Rectangle((int)Location.X, (int)Location.Y, (int)block.Width, (int)block.Height);
         }
 
         public void Update(GameTime gameTime)
